requirements.txt
gspread
pandas
oauth2client
arxiv
openai
python-dotenv
tweepy
nest_asyncio
beautifulsoup4
playwright
requests
pymupdf
tiktoken
Pillow